+++ controller/dto/article/article_like_dto.go
@@ -7,7 +7,7 @@ type CreateArticleLikeResponseDTO struct {
 	Title         string        `json:"title"`
 	Description   string        `json:"description"`
 	Body          string        `json:"body"`
-	FavoriteCount int           `json:"favorite_count"`
+	FavoriteCount int64         `json:"favorite_count"`
 	TagList       []string      `json:"tag_list,omitempty"`
 	IsFavorited   bool          `json:"is_favorited"`
 	Author        ArticleAuthor `json:"author"`
